implement own etnaviv drm driver to fix cache bug

+++ src/armsoc_exa.h
@@ -42,7 +42,7 @@
 struct ARMSOCEXABuf {
 	void *buf;
 	size_t size;
-	unsigned short pitch;
+	int pitch;
 	void *priv;
 };
